SPARK-130 moved manager to own file, moved config from project root
